Merge in KC/kopanocore from ~JENGELHARDT/kc:warn to master

* commit '227f3da57710af35f3dfd273f5c90c85cc45e6fa':
  build: cure compiler warning on ABEID_FIXED copying
  build: cure printf size warning for libfreebusy
  build: fix x32 compiler warning for printf formats
  build: fix compiler warning for ~0

+++ provider/server/ECThreadManager.cpp
@@ -150,9 +150,9 @@ static void log_request(struct soap *soap, int soaperr)
 		!st.imp.empty() ? bin2txt(st.imp).c_str() : "-",
 		st.func ?: "-", ers,
 		dur2dbl(st.sk_wall_dur), dur2dbl(st.enq_wall_dur),
-		dur2dbl(st.wi_wall_dur), static_cast<long>(st.wi_cpu[2].tv_sec), st.wi_cpu[2].tv_nsec / 1000,
-		dur2dbl(st.rh1_wall_dur), static_cast<long>(st.rh1_cpu[2].tv_sec), st.rh1_cpu[2].tv_nsec / 1000,
-		dur2dbl(st.rh2_wall_dur), static_cast<long>(st.rh2_cpu[2].tv_sec), st.rh2_cpu[2].tv_nsec / 1000,
+		dur2dbl(st.wi_wall_dur), static_cast<long>(st.wi_cpu[2].tv_sec), static_cast<long>(st.wi_cpu[2].tv_nsec / 1000),
+		dur2dbl(st.rh1_wall_dur), static_cast<long>(st.rh1_cpu[2].tv_sec), static_cast<long>(st.rh1_cpu[2].tv_nsec / 1000),
+		dur2dbl(st.rh2_wall_dur), static_cast<long>(st.rh2_cpu[2].tv_sec), static_cast<long>(st.rh2_cpu[2].tv_nsec / 1000),
 		st.agent.c_str()));
 }
